move db schema interfaces to their own file
