+++ src/pages/api/words.ts
@@ -3,7 +3,8 @@ import { jsonError, jsonResponse } from "../../server/responses";
 import { mkdir, writeFile } from "fs/promises";
 import { createWriteStream, ReadStream } from "fs";
 import { db, isDbError, Word } from "astro:db";
-import { wordFromDb, wordId, type WordType } from "../../../db/config";
+import { wordFromDb, wordId } from "../../../db/config";
+import { type WordType } from "../../../db/types";
 import { Marked } from "marked";
 import { baseUrl as markedBaseUrl } from "marked-base-url";
 import { markedEmoji } from "marked-emoji";
